change PATH in server tests to prefer blib/script over bin
Running the tests under a perlbrew environment may fail because the
system perl does not have the requisite features/modules installed. The
tests should normally be run after blib has been populated and so
putting that in front of the PATH in these tests ensures the right
interpreter is used.